% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Find the sum of two numbers.}
\usage{
add(x, y)
}
\arguments{
\item{x}{this is the first number}

\item{y}{this is the second number}
}
\description{
Find the sum of two numbers.
}
\examples{
add(2,6)
add (3,2)
}
