% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage.R
\name{calculate_percentage}
\alias{calculate_percentage}
\title{Calculate the percent in the Likert scale}
\usage{
calculate_percentage(data)
}
\arguments{
\item{data}{}
}
\description{
Calculate the percent in the Likert scale
}
\examples{
likert_data <- c(1, 2, 3, 3, 4, 5, 5, 5)
result <- calculate_percentage(likert_data)

}
