% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_answers.R
\name{grade_answers}
\alias{grade_answers}
\title{Dichotomize student's response}
\usage{
grade_answers(df, correct_answers)
}
\arguments{
\item{df}{input the students' response}

\item{correct_answers}{}
}
\value{
df dichotomized responses
}
\description{
Dichotomize student's response
}
\examples{
library(readxl)
students_df <- read_excel("students.xlsx")
correct_answers <- c("A", "B", "C", "D")
graded_df <- grade_answers(students_df, correct_answers)
data<-data.frame(graded_df)
data

}
